Remove isatty() detection
Second-guessing the user is rarely a good idea.  It is more important
for the user to be able to predict the behavior of a program than for
a program to predict the intent of the user.

Additionally, sys.stdin.isatty() == False does not actually mean that
an input file was piped in.  It could also mean, for example, that the
program is being run via an ssh command string or via a GUI
application or within some other script that does not pass along
stdin.

The failure mode here was dangerous: running the program on a single
filename argument in one of these ways would change a read from that
file into a write to that file.

Signed-off-by: Anders Kaseorg <andersk@mit.edu>